+++ updater/src/com/intellij/updater/DeleteAction.java
@@ -44,7 +44,11 @@ public class DeleteAction extends PatchAction {
 
   @Override
   protected void doApply(ZipFile patchFile, File backupDir, File toFile) throws IOException {
-    Utils.delete(toFile);
+    //NOTE: a folder can be deleted only in case if it does not contain any user's files/folders.
+    File[] listFiles = toFile.listFiles();
+    if (!toFile.isDirectory() || (listFiles != null && listFiles.length == 0)) {
+      Utils.delete(toFile);
+    }
   }
 
   @Override
